\name{ABACUS-package}
\alias{ABACUS-package}
\alias{ABACUS}
\docType{package}
\title{\packageTitle{ABACUS}}
\description{\packageDescription{ABACUS}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{ABACUS}
}
\author{
  \packageAuthor{ABACUS}
}
\references{
 	W. Zhang, D. Gilbert, D. Matteson. ABACUS: Unsupervised Multivariate Change Detection via Bayesian Source Separation. arXiv:1810.06167
}