% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accounts}
\alias{accounts}
\title{100 fake bank accounts}
\format{A data frame with 100 rows and 4 variables:
\describe{
  \item{owner}{random names}
  \item{iban}{random IBANs, don't necessarily make sense}
  \item{bic}{random BICs, don't necessarily make sense}
  \item{type}{a label for type}
}}
\usage{
accounts
}
\description{
This is used to create a test data base. IBAN and BIC do not necessarily make sense.
}
\keyword{datasets}
