% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_newDB.R
\name{Create_newDB}
\alias{Create_newDB}
\title{Create_newDB}
\usage{
Create_newDB(dbName = "my.db")
}
\arguments{
\item{dbName}{\code{chr} (\code{"my.db"}) name of database / file (and path)}
}
\value{
\code{TRUE} if successful
}
\description{
This function creates a new empty database.
}

