% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_testDB.R
\name{Create_testDB}
\alias{Create_testDB}
\title{Create_testDB}
\usage{
Create_testDB(dbPath = "~/data", dbName = "test")
}
\arguments{
\item{dbPath}{\code{chr} (\code{"~/data"}) path to database directory}

\item{dbName}{\code{chr} (\code{"test"}) name of database / file (\code{.db} will be added)}
}
\value{
\code{TRUE} if successful
}
\description{
This function creates a test database with 100 accounts and 1000 transactions.
}
\seealso{
Other test.db functions: \code{\link{FeatureExtraction}}
}

