% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CV}
\alias{CV}
\title{CV}
\usage{
CV(abt, feats, labs, k = 5, diagonal = FALSE)
}
\arguments{
\item{abt}{\code{num matrix} containing training data with rows as observations and columns as features}

\item{feats}{\code{data.frame} with columns \emph{name} and \emph{value} which identifies the features (columns)
of \code{abt} with \code{chr} values}

\item{labs}{\code{vector} defining class labels of rows in training data}

\item{diagonal}{\code{bool} (=\code{FALSE}) if true DDA instead of LDA is done}
}
\value{
\code{data.frame} with class labels and predictions by the classifier
}
\description{
This function trains a shrinkage discriminant analysis (sda) classifier as in \code{\link{Training}} in a \code{k}-fold 
cross validation.
A \code{data.frame} with class labels and predictions is returned which can be used for prediction power estimation.
}
\details{
This function uses the \emph{sda} package for training and ranking of a sda classifier.
Shrinkage intensity for correlation matrix, variances, and frequencies is estimated from the data.
With \code{diagonal} set to \code{TRUE} only the diagonal of the covariance matrix is used.
This speeds up the process and uses less memory.
}
\examples{
abt1 <- matrix(sample(0:1, 1000*100, replace = TRUE), 1000, 100)
feats1 <- data.frame(name = "test", value = 1:100)
labs1 <- sample(0:1, 1000, replace = TRUE)
err <- CV(abt1, feats1, labs1)
(acc <- sum(err$class == err$prediction) / nrow(err) * 100)

}
\seealso{
Other machine learning: \code{\link{Convert}},
  \code{\link{FeatureExtraction}},
  \code{\link{Prediction}}, \code{\link{Training}}
}

