% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicated.R
\name{Duplicated}
\alias{Duplicated}
\title{Duplicated}
\usage{
Duplicated(x, ...)
}
\description{
Generic function for finding duplicates.
\code{\link{Duplicated.Transactions}} is used to check whether transactions of a
\emph{Transactions} object (created with \code{\link{Predict.Transactions}})
are already present in the database.
}
\seealso{
Other procedures: \code{\link{Duplicated.Transactions}},
  \code{\link{Evaluate_Predictor}},
  \code{\link{Predict.Transactions}},
  \code{\link{Predict}}, \code{\link{Read.Transactions}},
  \code{\link{Read_csv}}, \code{\link{Read}},
  \code{\link{Update_Predictor}}
}

