% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enter.R
\name{Enter}
\alias{Enter}
\title{Enter}
\usage{
Enter(x, ...)
}
\description{
Generic function for inserting new rows in database.
\code{\link{Enter.Transactions}} is used with a
\emph{Transactions} object (created with \code{\link{Duplicated.Transactions}})
to insert transactions into the database.
}
\seealso{
Other procedures: \code{\link{Duplicated.Transactions}},
  \code{\link{Duplicated}},
  \code{\link{Enter.Transactions}},
  \code{\link{Evaluate_Predictor}},
  \code{\link{Predict.Transactions}},
  \code{\link{Predict}}, \code{\link{Read.Transactions}},
  \code{\link{Read_csv}}, \code{\link{Read}},
  \code{\link{Update_Predictor}}
}

