% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtraction.R
\name{FeatureExtraction}
\alias{FeatureExtraction}
\title{FeatureExtraction}
\usage{
FeatureExtraction(ta, pa)
}
\arguments{
\item{ta}{\code{data.frame} of \emph{transactions} joined with \emph{accounts} as returned 
by \code{\link{Select}("transactions", ...)}}

\item{pa}{\code{data.frame} of \emph{personalAccounts} joined with \emph{accounts} as returned 
by \code{\link{Select}("personalAccounts", ...)}}
}
\value{
\code{list} of 2 objects
\itemize{
   \item \emph{ABT} \code{int matrix} analytics base table containing features as columns, data points as rows
   \item \emph{FeatureList} \code{data.frame} description of \emph{ABT} columns with names and values
}
}
\description{
This function does a feature extraction to create an analytics base table from the \emph{transactions}
and the \emph{personalAccounts} tables.
The features are mostly binary (occurence of a word) or discrete (count of a word) and very sparse.
}
\details{
The analytics base table (\emph{ABT}) containes feature values as \code{int}.
Columns represent features, rows data points (= transactions).
The ABT is accompanied by a \emph{FeatureList} which describes the features.
It is necessary for conversion between different ABTs.
}
\examples{
Create_testDB("./db")
ta <- Select("transactions", "db/test.db")
pa <- Select("personalAccounts", "db/test.db")
res <- FeatureExtraction(ta, pa)

}
\seealso{
Other machine learning: \code{\link{CV}},
  \code{\link{Convert}}, \code{\link{Prediction}},
  \code{\link{Training}}
}

