% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insert.R
\name{Insert}
\alias{Insert}
\title{Insert}
\usage{
Insert(data, table, db, add_id = FALSE, enforce_foreign_keys = TRUE)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} data to be written into db table, without id (id will be auto incremented)}

\item{table}{\code{chr} of table name}

\item{db}{\code{chr} of database name/file from current directory}

\item{add_id}{\code{bool} (\code{FALSE}) whether 1st column named \emph{id} with values \code{NULL} should be added}

\item{enforce_foreign_keys}{\code{bool} (\code{TRUE}) whether to enforce rules on foreign keys}
}
\value{
chr HTML for creating ui elements.
}
\description{
Convenience function to write a \code{data.frame} or \code{matrix} into a database table.
The first column id will be added and incremented automatically if \code{add_id} is enabled.
}
\examples{
df <- data.frame(owner = "B. Clinton", iban = "IR98000020018267384", bic = "IR875TW78", type = "donations account")
Insert(df, "accounts", "db/mydb.db", add_id = TRUE)

}

