% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intersect.R
\name{Intersect}
\alias{Intersect}
\title{Intersect}
\usage{
Intersect(data, table, db)
}
\arguments{
\item{data}{\code{data.frame} whose entries are compared to a database table}

\item{table}{\code{chr} of database table name}

\item{db}{\code{chr} full file name with path of database}
}
\value{
\code{bool} vector or list of \code{bool} vector and a \code{data.frame}
}
\description{
Convenience function for comparing a \code{data.frame} with table entries of a database.
This is used to see if certain entries already exist in the database.
}
\details{
What columns exactly are compared depends on the \code{table} value.
\itemize{
   \item \emph{accounts} the table is searched for equal values in both columns \emph{bic} and \emph{iban}.
         A \code{bool} vector indicating for each row in \code{data} whether it exists in \code{table}.
   \item \emph{storage} or \emph{personalAccounts} the table is searched for equal values in columns \emph{name} and
         \emph{type} respectively. A \code{bool} vector indicating for each row in \code{data} whether it exists in \code{table}.
}
}
\examples{
db <- "db/test.db"
Create_testDB(db)
df <- Select("transactions", "db/test.db", le = list(payor_id = 2), eq = list(type = c("food", "purchase")))

}
\seealso{
Other SQLite handler functions: \code{\link{InsertBLOB}},
  \code{\link{Insert}}, \code{\link{SelectBLOB}},
  \code{\link{Select}}, \code{\link{UpdateBLOB}},
  \code{\link{Update}}
}

