% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{personalAccounts}
\alias{personalAccounts}
\title{2 fake personal bank accounts}
\format{A data frame with 2 rows and 2 variables:
\describe{
  \item{account}{id of personal account, references \emph{accounts}}
  \item{comment}{purpose, nature of account}
}}
\usage{
personalAccounts
}
\description{
This is used to create a test data base. They reference the table \emph{accounts}.
}
\keyword{datasets}

