% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{Predict}
\alias{Predict}
\title{Predict}
\usage{
Predict(x, ...)
}
\description{
Generic function for making a prediction with certain objects.
\code{\link{Predict.Transactions}} is used to predict transaction types of a 
\emph{Transactions} object (created with \code{\link{Read.Transactions}} for example).
}
\seealso{
Other procedures: \code{\link{Evaluate_Predictor}},
  \code{\link{Predict.Transactions}},
  \code{\link{Read.Transactions}}, \code{\link{Read_csv}},
  \code{\link{Read}}, \code{\link{Update_Predictor}}
}

