% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{Predict.Transactions}
\alias{Predict.Transactions}
\title{Predict.Transactions}
\usage{
\method{Predict}{Transactions}(x)
}
\arguments{
\item{x}{\code{Transactions} object (created with \code{\link{Read.Transactions}} for example)}
}
\value{
\code{Transactions} object, a list of 4 elements:
\itemize{
    \item \emph{Transactions} a data.frame of the file that was read
    \item \emph{Reference} a data.frame of the reference account
    \item \emph{db} character of database used
    \item \emph{Prediction} a data.frame of transactions with additional columns of account information.
          It includes a column with the predicted type of each transaction.
}
}
\description{
This function uses a machine learning method to classify each type of transaction for a \emph{Transactions} object
as created with \code{\link{Read.Transactions}} for example.
A table with the actual transactions and the predicted types is returned.
}
\details{
\enumerate{
    \item If there are \emph{new accounts} (not existent in database yet), these accounts are written into the database.
    Also, if the user did make changes to these accounts, these changes are applied to the actual transactions
    accordingly.
    \item A complete \emph{transactions} table is created by combining the original transactions table with all
    information about the accounts. Numeric values are converted to Integers in this step (* 100).
    \item If a \emph{Prediction Model} was already created, it is read from the database
    \item \emph{Feature Extraction} and \emph{Prediction} are done using the model
    
}
}
\examples{
db <- "test.db"
Create_testDB(db)

f <- system.file("extdata", "test_transactions.csv", package = "abacus")
cols <- list(name = 6, iban = 7, bic = 8, date = 3, reference = 5, entry = 4, value = 9, currency = 10)
tas <- Read_csv("giro", f, cols, db)
tas <- Read(tas)

params <- list(nFeats = 200, DDL = FALSE, time = list(start = as.Date("2010-1-1"), end = as.Date("2011-1-1")))
InsertBLOB("Params", params, db)
Update_Predictor(db)

pred <- Predict(tas)

}
\seealso{
Other procedures: \code{\link{Duplicated.Transactions}},
  \code{\link{Duplicated}},
  \code{\link{Evaluate_Predictor}}, \code{\link{Predict}},
  \code{\link{Read.Transactions}}, \code{\link{Read_csv}},
  \code{\link{Read}}, \code{\link{Update_Predictor}}
}

