% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{Read.Transactions}
\alias{Read.Transactions}
\title{Read}
\usage{
\method{Read}{Transactions}(x)
}
\arguments{
\item{x}{\code{Transactions} object 
(created with \code{\link{Read_csv}} for example)}
}
\value{
\code{Transactions} object, a list of 4 elements:
\itemize{
    \item \emph{Transactions} a data.frame of the file that was read
    \item \emph{Reference} a data.frame of the reference account
    \item \emph{db} character of database used
    \item \emph{NewAccounts} a data.frame of accounts that appear in 
    \emph{Transactions}, but not in the database
}
}
\description{
This procedure identifies accounts occuring in a \emph{Transactions} object 
which cannot be found in the associated database.
It then adds these accounts to the \emph{Transactions} object and returns it.
}
\examples{
db <- "test.db"
Create_testDB(db)

f <- system.file("extdata", "test_transactions.csv", package = "abacus")
cols <- list(name = 6, iban = 7, bic = 8, date = 3, 
             reference = 5, entry = 4, value = 9, currency = 10)
tas <- Read_csv("giro", f, cols, db)
tas <- Read(tas)

}
\seealso{
Other procedures: \code{\link{Duplicated.Transactions}},
  \code{\link{Duplicated}},
  \code{\link{Enter.Transactions}}, \code{\link{Enter}},
  \code{\link{Evaluate_Predictor}},
  \code{\link{Predict.Transactions}},
  \code{\link{Predict}}, \code{\link{Read_csv}},
  \code{\link{Read}}, \code{\link{Update_Predictor}}
}
