% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_csv.R
\name{Read_csv}
\alias{Read_csv}
\title{Read_csv}
\usage{
Read_csv(ref, fileName, columns, db, head = TRUE, colSep = "\\t",
  decSep = ",", quoteChar = "\\"", commentChar = "", nSkip = 0,
  nMax = -1, dateFormat = "\%d.\%m.\%Y")
}
\arguments{
\item{ref}{\code{chr} for the \emph{type} of the reference account of the uploaded file}

\item{fileName}{\code{chr} for the file to be read (should be a \emph{csv})}

\item{columns}{\code{list} which defines which columns of the read table (from file) contain which information:
It must specify \emph{name}, \emph{iban}, \emph{bic}, \emph{date}, 
\emph{reference}, \emph{entry}, \emph{value}, \emph{currency} as list elements in this order,
with their according columns in the table as \code{int} values.}

\item{db}{\code{chr} the database used / file name and path of database}

\item{head}{\code{bool} (=\code{TRUE}) whether the first line of the file defines column names of the table}

\item{colSep}{\code{chr} (=\code{"\\t"}) seperator for columns in the table}

\item{decSep}{\code{chr} (=\code{","}) decimal seperator for numbers}

\item{quoteChar}{\code{chr} (=\code{"\\""}) symbol used for quotes in table}

\item{commentChar}{\code{chr} (=\code{""}) symbol introducing comments in the table}

\item{nSkip}{\code{int} (=\code{0}) how many lines to skip when reading the file (from top to bottom)}

\item{nMax}{\code{int} (=\code{-1}) maximum number of lines the read (negative to read all lines of file)}

\item{dateFormat}{\code{chr} (=\code{"\%d.\%m.\%Y"} for"13.03.1990") date format to use for converting the date column}
}
\value{
\code{Transactions} object which is a list of 3 elements:
\itemize{
    \item \emph{Transactions} a data.frame of the file that was read
    \item \emph{Reference} a data.frame of the reference account
    \item \emph{db} character of database used
}
}
\description{
This function is used to read a \emph{csv} file and create a \emph{Transactions} object from it.
This object contains the transactions in a standardized form and information about the reference account.
}
\details{
\enumerate{
 \item \emph{reference account} information is retrieved from the database
 \item \emph{file} is read as a table as specified in arguments
 \item table is \emph{converted} as specified with arguments \code{columns} and \code{dateFormat}
 \item \emph{Transactions} object is returned
}
}
\examples{
db <- "test.db"
Create_testDB(db)

f <- system.file("extdata", "test_transactions.csv", package = "abacus")
cols <- list(name = 6, iban = 7, bic = 8, date = 3, reference = 5, entry = 4, value = 9, currency = 10)
tas <- Read_csv("giro", f, cols, db)

}
\seealso{
Other procedures: \code{\link{Duplicated.Transactions}},
  \code{\link{Duplicated}},
  \code{\link{Evaluate_Predictor}},
  \code{\link{Predict.Transactions}},
  \code{\link{Predict}}, \code{\link{Read.Transactions}},
  \code{\link{Read}}, \code{\link{Update_Predictor}}
}

