% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.R
\name{Select}
\alias{Select}
\title{Select}
\usage{
Select(table, db, eq = NULL, ge = NULL, le = NULL,
  enforce_foreign_keys = TRUE, check_query = FALSE)
}
\arguments{
\item{table}{\code{chr} of table name, will automatically be \code{INNER JOIN}ed in case relation exists}

\item{db}{\code{chr} full file name with path of database}

\item{eq}{\code{list} (\code{NULL}) defining a condition with \code{=} (equal). . 
Element name specifies column, value its value (see details)}

\item{ge}{\code{list} (\code{NULL}) defining a condition with \code{>=} (greater-equal). 
Element name specifies column, value its value (see details)}

\item{le}{\code{list} (\code{NULL}) defining a condition with \code{<=} (lesser-equal). . 
Element name specifies column, value its value (see details)}

\item{enforce_foreign_keys}{\code{bool} (\code{TRUE}) whether to enforce rules on foreign keys}

\item{check_query}{\code{bool} (\code{FALSE}) whether to just return the SQL query without actually sending it}
}
\value{
\code{data.frame} of table
}
\description{
Convenience function for \code{SELECT} queries. 
One of the 6 table names can be given with optional \code{WHERE} statements.
In case transactions, capital or personalAccounts is selected, a \code{INNER JOIN} with accounts will be done (2 joins for transactions).
}
\details{
A \code{WHERE} condition can be added with arguments \code{eq} (equal), \code{ge} (greater-equal), \code{le} (lesser-equal).
Conditions are given as list for the desired relation. The name of a list element defines a column and its value the value.
Multiple \code{ge} and \code{se} conditions are combined by \code{AND} and \code{eq} conditions by \code{OR}.

You can set \code{check_query} to \code{TRUE} to see the query.
}
\examples{
Create_testDB("./db")
df <- Select("transactions", "db/test.db", le = list(payor_id = 2), eq = list(type = c("food", "purchase")))

}
\seealso{
Other SQLite handler functions: \code{\link{InsertBLOB}},
  \code{\link{Insert}}, \code{\link{SelectBLOB}}
}

