% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.R
\name{SelectBLOB}
\alias{SelectBLOB}
\title{SelectBLOB}
\usage{
SelectBLOB(name, db, table = "storage")
}
\arguments{
\item{name}{\code{chr} name of BLOB}

\item{db}{\code{chr} full file name with path of database}

\item{table}{\code{chr} (="storage") name of the storage table
in db}
}
\value{
\code{R} object selected
}
\description{
Convenience function to retrieve any kind of BLOB stored in table 
\emph{storage} of SQLite database
}
\examples{
db <- "db/test.db"
Create_testDB(db)
x <- list(a = 1:5, b = list(c = c("a", "b")))
InsertBLOB("test2", x, db)
SelectBLOB("test2", db)

}
\seealso{
Other SQLite handler functions: \code{\link{InsertBLOB}},
  \code{\link{Insert}}, \code{\link{Intersect}},
  \code{\link{Select}}, \code{\link{UpdateBLOB}},
  \code{\link{Update}}
}

