% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{Table}
\alias{Table}
\title{Table}
\usage{
Table(data, colNames = NULL, bRownames = FALSE, style = "default",
  class = "display", dom = "flrtip", ordering = NULL,
  alignment = list(centre = NULL, justify = NULL, left = NULL),
  formatCurr = NULL, formatPerc = NULL, formatRoun = NULL,
  bButtons = FALSE, bResponsive = FALSE, pageLen = 15, bScroll = FALSE,
  filename = "*", esc = TRUE)
}
\arguments{
\item{data}{\code{dataframe} to be rendered as table}

\item{colNames}{\code{NULL} or \code{chr} arr of same length as rows in data. if not \code{NULL} \code{chr arr} is
taken as column names instead of column names of data}

\item{bRownames}{\code{bool} whether to show rownames}

\item{style}{\code{chr} defining style. currently there is only \code{"default"} and \code{"bootstrap"}}

\item{class}{\code{chr} defining class. \code{"display"} is default, \code{"stripe hover"} is also nice
there are many stlye combinations possible (DT website)}

\item{dom}{\code{chr} for dom arguments. define table elements in order. \code{f} filtering, 
\code{r} processing, \code{t} table, \code{i} information, \code{p} pagination, \code{"frtip"} is default}

\item{ordering}{\code{NULL} or \code{list} of \code{num arr} and \code{chr 'asc'} or \code{'desc'}. e.g. \code{list(4, 'asc')}}

\item{alignment}{\code{list} of 3 elements \emph{centre, justify}, and \emph{left}. they can each be \code{NULL}(default) or
a \code{num} arr for which columns to be aligned accordingly. other columnss are right aligned}

\item{formatCurr}{\code{NULL} or \code{list} of 2 elements \emph{cols} (\code{num arr}) and \emph{curr} (\code{chr}) for 
currency formatting of \emph{cols} columns}

\item{formatPerc}{\code{NULL} or \code{list} of 1 element \emph{cols} (\code{num arr}) for percentage formatting of cols columns}

\item{formatRoun}{\code{NULL} or \code{list} of 2 elements \emph{cols} (\code{num arr}) and \emph{digits} (\code{num}) for rounding numbers
of cols columns to number of digits}

\item{bResponsive}{\code{bool} whether responsive is activated: if table is to narrow, columns are excluded
and there is a ugly plus symbol where they can be displayed by clicking}

\item{pageLen}{\code{num} for how many rows are displayed}

\item{bScroll}{\code{bool} whether xScroll is enabled (horizontal scrolling)}

\item{filename}{\code{chr} filename for downloaded data}

\item{esc}{\code{bool} whether to escape HTML, if not JS callback ensures input bindings}

\item{buttons}{\code{NULL} or \code{chr} arr defining download buttons used in table ('copy', 'csv', 'excel', 'pdf', 'print')}
}
\value{
\code{datatables}, \code{htmlwidget} object
}
\description{
Convenience function for rendering DT datatables.
}
\examples{
Table_DT(iris, bButtons = TRUE, alignment = list(left = 2), dom = "t")

}

