% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transactions}
\alias{transactions}
\title{1000 fake transactions}
\format{A data frame with 1000 rows and 7 variables:
\describe{
  \item{payor}{id of payor account, references \emph{accounts}}
  \item{payee}{id of payee account, references \emph{accounts}}
  \item{date}{date of transaction}
  \item{reference}{reference text}
  \item{entry}{booking entry}
  \item{value}{in cents (decimal currency)}
  \item{currency}{currency code}
}}
\usage{
transactions
}
\description{
This is used to create a test data base. They reference the table \emph{accounts}.
}
\keyword{datasets}

