% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update.R
\name{Update}
\alias{Update}
\title{Update}
\usage{
Update(data, table, check, db, enforce_foreign_keys = TRUE)
}
\arguments{
\item{data}{\code{data.frame} of data to be written into db 
table}

\item{table}{\code{chr} of table name}

\item{check}{\code{chr} of column name(s) for which updated 
rows have to have equal content}

\item{db}{\code{chr} full file name with path of database}

\item{enforce_foreign_keys}{\code{bool} (\code{TRUE}) whether to enforce 
rules on foreign keys}
}
\value{
\code{TRUE} if successful
}
\description{
Convenience function to update tables of a SQLite database.
For each row of the updates (\code{data}), the function looks for rows in the
database table which have the same values as \code{data} for the columns 
defined in \code{check}. These rows are updated with the information of that 
\code{data} row.
}
\examples{
db <- "db/test.db"
Create_testDB(db)
df <- data.frame(
  id = 3:4, 
  owner = rep("test",2), 
  iban = rep("test",2), 
  bic = rep("test",2), 
  stringsAsFactors = FALSE
 )
Update(df, "accounts", "id", db)
Select("accounts", db, eq = list(id = c(3,4)))

}
\seealso{
Other SQLite handler functions: \code{\link{InsertBLOB}},
  \code{\link{Insert}}, \code{\link{Intersect}},
  \code{\link{SelectBLOB}}, \code{\link{Select}},
  \code{\link{UpdateBLOB}}
}

