% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update.R
\name{UpdateBLOB}
\alias{UpdateBLOB}
\title{UpdateBLOB}
\usage{
UpdateBLOB(name, data, db, table = "storage")
}
\arguments{
\item{name}{\code{chr} name of BLOB}

\item{data}{any object that is to be stored as BLOB}

\item{db}{\code{chr} full file name with path of database}

\item{table}{\code{chr} table name ("storage")}
}
\value{
\code{TRUE} if successful
}
\description{
Convenience function to write any kind of object into a SQLite table.
}
\examples{
db <- "db/test.db"
Create_testDB(db)
d <- list(test = "some test")
InsertBLOB("test3", list(a = 1:10, b = list(a = letters)), db)
UpdateBLOB("test3", d, db)
d2 <- SelectBLOB("test3", db)

}
\seealso{
Other SQLite handler functions: \code{\link{InsertBLOB}},
  \code{\link{Insert}}, \code{\link{Intersect}},
  \code{\link{SelectBLOB}}, \code{\link{Select}},
  \code{\link{Update}}
}

