% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadSettings.R
\name{UploadSettings}
\alias{UploadSettings}
\title{UploadSettings}
\usage{
UploadSettings(input, output, session, db)
}
\arguments{
\item{input}{\code{list} from shiny session}

\item{output}{\code{list} from shiny session}

\item{session}{\code{list} from shiny session}

\item{db}{\code{chr} of database file (full path and file name)}
}
\value{
\code{TRUE}
}
\description{
UploadModal is a modal with multiple pages for entering transactions into the database.
}
\details{
This is the server logic of the modal.
While the user goas through 4 pages in the modal following is done.
\enumerate{
   \item \code{\link{Read}} method extracts new accountss from \emph{transactions} object.
   \item \code{\link{Predict}} method enters new accounts into the database and 
   then predicts each transaction.
   \item \code{\link{Duplicated}} reads the database to see if transactions were already entered.
   \item \code{\link{Enter}} finally inserts new transaction into the database.
}
}
\examples{


}
\seealso{
Other application functions: \code{\link{UploadModalUI}},
  \code{\link{UploadModal}},
  \code{\link{UploadSettingsUI}}, \code{\link{abacusApp}}
}

