\name{ABACUS-package}
\alias{ABACUS-package}
\docType{package}

\title{
  Apps Based Activities for Communicating and Understanding Statistics
}

\description{
  Using Shiny-based apps, ABACUS demonstrates different statistical concepts on
  sampling and hypothesis testing within a browser environment.
  It integrates the simulation approach with the graphical presentation to present different statistical tools.
  The user can alter \strong{Simulation Feature}, \strong{Population Parameters},
  \strong{Sample Characteristics} and \strong{Distribution Function}
  to explore the influence of these changes on sampling and hypothesis testing.
}

\details{

  \tabular{ll}{
  Package: \tab ABACUS\cr
  Type: \tab Package\cr
  Version: \tab 1.0.0\cr
  Date: \tab 2019-07-31\cr
  License: \tab GPL-3 \cr
  }

  ABACUS uses simulation to explore the statistical principles. There is no need to upload any data.
  The order of ABACUS functions that can be included in a classroom environment are:
  \code{\link{shiny_dnorm}}, \code{\link{shiny_sampling}},
  \code{\link{shiny_onesampz}}, \code{\link{shiny_onesampt}},
  \code{\link{shiny_dnorm_dt}},
  \code{\link{shiny_twosampt}} and \code{\link{shiny_anova}}.

}


\note{
  \url{https://shiny.abdn.ac.uk/Stats/apps/}
}


\author{
  Mintu Nath<dr.m.nath@gmail.com>
}


\keyword{ package }


\seealso{
  \code{\link{shiny_dnorm}}, \code{\link{shiny_sampling}},
  \code{\link{shiny_onesampz}}, \code{\link{shiny_onesampt}},
  \code{\link{shiny_twosampt}} and \code{\link{shiny_anova}}
}



