% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abacusApp.R
\name{abacusApp}
\alias{abacusApp}
\title{abacusApp}
\usage{
abacusApp(db)
}
\arguments{
\item{db}{\code{chr} database file (full path with file name)}
}
\value{

}
\description{
The App
}
\examples{


}
\seealso{
Other application functions: \code{\link{UploadModalUI}},
  \code{\link{UploadModal}},
  \code{\link{UploadSettingsUI}},
  \code{\link{UploadSettings}}
}

