% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_testDB.R
\name{Create_testDB}
\alias{Create_testDB}
\title{Create_testDB}
\usage{
Create_testDB(dbName = "test.db")
}
\arguments{
\item{dbName}{\code{chr} (\code{"test.db"}) name of database /
file (in current working dir)}
}
\value{
\code{TRUE} if successful
}
\description{
This function creates a test database with 100 accounts and 1000 
transactions for trying out stuff.
}
