% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CV}
\alias{CV}
\title{CV}
\usage{
CV(abt, feats, labs, k = 5, ...)
}
\arguments{
\item{abt}{\code{num matrix} containing training data with rows 
as observations and columns as features}

\item{feats}{\code{data.frame} with columns \emph{name} and 
\emph{value} which identifies the features (columns)
of \code{abt} with \code{chr} values}

\item{labs}{\code{vector} defining class labels of rows in 
training data}

\item{...}{arguments passed to \code{link{Training}}}
}
\value{
\code{data.frame} with class labels and predictions by the classifier
}
\description{
This function trains a shrinkage discriminant analysis (sda) classifier as 
in \code{\link{Training}} in a \code{k}-fold 
cross validation.
A \code{data.frame} with class labels and predictions is returned which can 
be used for prediction power estimation.
}
\examples{
abt1 <- matrix(sample(0:1, 1000*100, replace = TRUE), 1000, 100)
feats1 <- data.frame(name = "test", value = 1:100)
labs1 <- sample(0:1, 1000, replace = TRUE)
err <- CV(abt1, feats1, labs1)
(acc <- sum(err$class == err$prediction) / nrow(err) * 100)

}
\seealso{
Other machine learning: \code{\link{Convert}},
  \code{\link{FeatureExtraction}},
  \code{\link{Prediction}}, \code{\link{Training}}
}
