% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insert.R
\name{Insert}
\alias{Insert}
\title{Insert}
\usage{
Insert(data, table, db, add_id = FALSE, enforce_foreign_keys = TRUE)
}
\arguments{
\item{data}{\code{data.frame} of data to be written into db 
table}

\item{table}{\code{chr} of table name}

\item{db}{\code{chr} full file name with path of database}

\item{add_id}{\code{bool} (\code{FALSE}) whether 1st column 
named \emph{id} with values \code{NULL} should 
be added (will be autoincremented in table 
\emph{accounts})}

\item{enforce_foreign_keys}{\code{bool} (\code{TRUE}) whether to enforce 
rules on foreign keys}
}
\value{
\code{TRUE} if successful
}
\description{
Convenience function to write a \code{data.frame} or \code{matrix} into a 
database table. The first column id will be added and incremented 
automatically if \code{add_id} is enabled.
}
\examples{
db <- "db/test.db"
Create_testDB(db)
df <- data.frame(
  owner = "B. Clinton", 
  iban = "IR98000020018267384", 
  bic = "IR875TW78", 
  type = "donations account"
)
Insert(df, "accounts", "db/test.db", add_id = TRUE)

}
\seealso{
Other SQLite handler functions: \code{\link{InsertBLOB}},
  \code{\link{Intersect}}, \code{\link{SelectBLOB}},
  \code{\link{Select}}, \code{\link{UpdateBLOB}},
  \code{\link{Update}}
}

