% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insert.R
\name{InsertBLOB}
\alias{InsertBLOB}
\title{InsertBLOB}
\usage{
InsertBLOB(name, data, db, table = "storage")
}
\arguments{
\item{name}{\code{chr} name of BLOB}

\item{data}{any object that is to be stored as BLOB}

\item{db}{\code{chr} full file name with path of database}

\item{table}{\code{chr} table name ("storage")}
}
\value{
\code{TRUE} if successful
}
\description{
Convenience function to write any kind of object into a SQLite table.
}
\examples{
Insert("test", list(a = 1:10, b = mtcars), "db/mydb.db")

}
\seealso{
Other SQLite handler functions: \code{\link{Insert}}
}

